Add kleine Verbesserung während der Übung

+++ 03-PDE/pde/partdiff-seq.c
@@ -153,7 +153,7 @@ double getResiduum(double h, struct options *options, int x, int y, double star)
 }
 
 double f(double xh, double yh) {
-    return TWO_PI_SQUARE * sin(yh * PI) * sin(xh * PI);
+    return (double) TWO_PI_SQUARE * sin(yh * PI) * sin(xh * PI);
 }
 
 /* ************************************************************************ */
